feat(Scheduled Jobs): implement automated tasks for expired reservations
- Added SchedulingConfig with @EnableScheduling.
- Added ReservationCleanupJob to automatically mark PENDING reservations as EXPIRED if their start time has passed.
- Added ReservationStatus.EXPIRED.
- Changed Reservation.totalPrice from Double to BigDecimal to correctly map to PostgreSQL numeric schema and prevent validation errors.
- Refactored FootballAppApplication to com.huseyinsacikay to fix component scanning without workaround annotations.
- Implemented ReservationCleanupJobTest to ensure scheduled job behaves correctly.
- Checked off Phase 08 in ROADMAP.md.

+++ backend/src/main/java/com/huseyinsacikay/entity/Reservation.java
@@ -2,6 +2,8 @@ package com.huseyinsacikay.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -30,7 +32,7 @@ public class Reservation {
     private LocalDateTime endTime;
 
     @Column(name = "total_price")
-    private Double totalPrice;
+    private BigDecimal totalPrice;
 
     @Enumerated(EnumType.STRING)
     @Builder.Default
